changes to respect ISO C99 (tested with gcc under Ubuntu 9)
git-svn-id: http://beaengine.googlecode.com/svn/trunk@52 e67b0f8a-f93e-11de-ba51-6b4196a2b81e

+++ beaengineSources/Includes/Routines_Disasm.c
@@ -484,10 +484,10 @@ void __bea_callspec__ CalculateRelativeAddress(UInt64 * pMyAddress, Int64 MyNumb
 {
     RelativeAddress = 1;
     if (EIP_VA != 0) {
-        *pMyAddress = EIP_VA+(Int64) MyNumber;
+        *pMyAddress = (UInt64) (EIP_VA+(Int64) MyNumber);
     }
     else {
-        *pMyAddress = EIP_REAL+(Int64) MyNumber;
+        *pMyAddress = (UInt64) (EIP_REAL+(Int64) MyNumber);
     }
 }
